+++ testing/tests/DevExpress.ui.widgets/fileManagerParts/arrayProvider.tests.js
@@ -3,7 +3,7 @@ const { test } = QUnit;
 import 'ui/file_manager';
 import ObjectFileSystemProvider from 'file_management/object_provider';
 import FileSystemItem from 'file_management/file_system_item';
-import ErrorCode from 'file_management/errors';
+import ErrorCode from 'file_management/file_system_error_code';
 import { fileSaver } from 'exporter/file_saver';
 
 import { createUploaderFiles, createUploadInfo } from '../../../helpers/fileManagerHelpers.js';
